Iss #160: Introduced a unified Dockerfile with multiple targets.
Update Feb 5th: Dropping changes to pyproject.toml as we revert back to having individual Dockerfiles soon.

We can leverage the same Dockerfile to build the images for the different auth components.

Deleted the old Dockerfiles and requirements files.

<Outdated> Any requirements from the old Dockerfile were moved to the pyproject.toml.

Introduced an overview README in the top-level auth directory that points to the READMEs for each component.

README may not to be re-introduced because of conflicts during a recent rebase with main branch.

Signed-off-by: Gloire Rubambiza <gloire@ibm.com>